requirements.txt
fastmcp
pyautogui
Pillow==10.4.0
